github: add issue type as link to podman-desktop
I see a lot of podman-desktop bug reports on the podman repo. This is
not the correct place to file these reports but many users seem to be
unaware of that. This is an attempt to point them in the right
direction. It should show up as extra entry in the issue type
selection so hopefully users while read that and follow the link in the
desktop repository.

See this link for the github docs about the configuration:
https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/configuring-issue-templates-for-your-repository#configuring-the-template-chooser

Signed-off-by: Paul Holzinger <pholzing@redhat.com>