Fix flakey machine starts
When starting a podman machine, we rely on SSH-ing into the machine in order to mount volumes. This requires the the machine to be running and listening to SSH connections. Just being able to connect to the SSH port is not a sufficient enough check to ensure the machine is listening to connections - we now check that the connection returns a string - which openSSH does by default on opening a connection.

[NO NEW TESTS NEEDED]

Signed-off-by: Ashley Cui <acui@redhat.com>