e2e: default to "podman version" in is-server-up tests
Every e2e test starts its own podman server. And every one of
those tests runs 'podman info' to check that the server is
up and running. This makes test logs painful to read, because
you click a link, get taken to the test name, then have to
scroll down and down past 170+ lines to get to the useful
part of the log.

Workaround: run '...info --format SOMETHING', so we just get
one short line of output. The SOMETHING is irrelevant, none
of it is actually interesting for default testing.

As a special case, on tests that run RestartRemoteService()
(containers.conf and network config changes), display the full
'podman info' output. This may prove useful some day. It also
makes it very obvious to even the most casual reader that
something different is happening.

Also, reduce retry timeout (2s -> .25s). The retry doesn't
seem to trigger often, but when it does, 2s is excessive.

Signed-off-by: Ed Santiago <santiago@redhat.com>