Add bash-completion for podman inspect and default to container-type
The command:
```podman inspect <ConatinerID> -f "{{."```
was not giving any output in console for bash-completion but the same was working fine for the command:
```podman container inspect <ConatinerID> -f "{{."```
So added a flow to handle the former command by defaulting to type CONTAINER and restrict the flow ofbash-completion if other types (eg. IMAGE)  is encountered.

Signed-off-by: Chetan Giradkar <cgiradka@redhat.com>