Fix bug report issue template README link
The relative link seems flakey: depending where you're creating the issue this may not take you to the right place.

For example, from https://github.com/containers/podman/issues/new it takes you to https://github.com/containers/podman/README.md, which does not exist.

To fix this, I've replaced it with an absolute link to the README for people to find the version easily. Other places in the same issue template use absolute paths to places in the repo e.g. the wiki, so I think this is acceptable.

Signed-off-by: Adam Jones <domdomegg+git@gmail.com>