Allow rootless containers to use AppArmor profiles
These changes make it possible to specify an AppArmor profile to run a
container under via the --security-opt apparmor=... flag when rootless.
Podman previously required root privileges to specify an AppArmor in
order to check /sys/kernel/security/apparmor/profiles to see whether or
not the provided AppArmor profile was loaded into the kernel. After
these changes, Podman will instead error out when it attempts to write
/proc/thread-self/attr/exec with a non-existent profile.

Note that Podman can't use apparmor_parser to load a new profile into
the kernel without root. There isn't a convenient way to allow
generateSpec to run CheckProfileAndLoadDefault rootlessly from
containers/common in its current implementation, so I've removed all
uses of that function from generateSpec (even when we're running as
root). In theory this would mean that under these changes we won't be
able to load a default profile for containers. However, it turns out
that in practice there was never any path through the code where Podman
would load the default profile anyways, so this is a non-breaking
change.

Fixes an issue previously described in:
https://github.com/containers/common/issues/958

Signed-off-by: Will Shand <wss2ec@g.ucla.edu>