Makefile: add warning about RHEL8 failing man pages
The new man-page table check (#19278) fails on RHEL8 because
go-md2man is 2.00 and we require 2.02.

After much deliberation.... just ignore the problem. Add a
bold note to that part of the Makefile, and hope that when
the test fails on the next RHEL branching someone will look
at the Makefile, find the comment, and disable that test.

Got any better ideas?

Signed-off-by: Ed Santiago <santiago@redhat.com>